we now use off_t for the sizes (sorry Ingo for bugging :) )
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@14538 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/apps/installer/CopyEngine.cpp
@@ -10,7 +10,7 @@
 #include <DiskDeviceTypes.h>
 #include <Path.h>
 
-extern void SizeAsString(int32 size, char *string);
+extern void SizeAsString(off_t size, char *string);
 
 class SourceVisitor : public BDiskDeviceVisitor 
 {
